# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Regalia KMS API",
    "version": "1.0.0",
    "description": "Purpose-shaped hardware-backed cryptographic operations. Version 1 is a contract; no production implementation is implied."
  },
  "servers": [
    {
      "url": "https://kms.example.invalid"
    }
  ],
  "security": [
    {
      "mutualTLS": []
    }
  ],
...
  "paths": {
...
    "/v1/operations/certificate-sign": {
      "post": {
        "operationId": "certificateSign",
        "tags": [
          "operations"
        ],
        "parameters": [
          {
            "name": "X-Request-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "description": "Must exactly equal context.nonce; every value is accepted at most once.",
            "schema": {
              "type": "string",
              "minLength": 16,
              "maxLength": 128
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnwrapRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authorized unwrapped data key.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/Denied"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/InvalidOperation"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        },
        "summary": "Sign a certificate request",
        "description": "Issues a certificate from a PKCS#10 CSR in DER, at most 8 KiB. The CSR is a request, not an instruction: the issuer profile decides the namespace, lifetime and key usage."
      }
    },
    "/v1/operations/key-agreement": {
      "post": {
        "operationId": "keyAgreement",
        "tags": [
          "operations"
        ],
        "parameters": [
          {
            "name": "X-Request-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "description": "Must exactly equal context.nonce; every value is accepted at most once.",
            "schema": {
              "type": "string",
              "minLength": 16,
              "maxLength": 128
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnwrapRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authorized unwrapped data key.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/Denied"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/InvalidOperation"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        },
        "summary": "Derive a shared key",
        "description": "Performs ECDH against a peer public key in PKIX DER, at most 4 KiB, and returns a 32-byte key derived through HKDF-SHA256 bound to the request context. The raw shared secret is never returned."
      }
    },
    "/v1/operations/release-secret": {
      "post": {
        "operationId": "releaseSecret",
        "tags": [
          "operations"
        ],
        "parameters": [
          {
            "name": "X-Request-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "description": "Must exactly equal context.nonce; every value is accepted at most once.",
            "schema": {
              "type": "string",
              "minLength": 16,
              "maxLength": 128
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReleaseSecretRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authorized secret plaintext.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/Denied"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/InvalidOperation"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        },
        "summary": "Release an enveloped secret",
        "description": "Unwraps a regalia-envelope-v2 envelope of at most 64 KiB and returns the secret plaintext. The envelope must belong to the object the route authorised."
      }
    },
    "/v1/operations/seal-envelope": {
      "post": {
        "operationId": "sealEnvelope",
        "tags": [
          "operations"
        ],
        "summary": "Produce a regalia-envelope-v2 document for a secret the caller encrypted locally.",
        "description": "The request carries the assembled parts (ciphertext, nonce, data key) and never the plaintext. The KMS proves possession of the data key by a local AEAD round-trip, then wraps the key against the object's commissioned KEK. The response is the envelope document; the KMS stores no copy.",
        "parameters": [
          {
            "name": "X-Request-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "description": "Must exactly equal context.nonce; every value is accepted at most once.",
            "schema": {
              "type": "string",
              "minLength": 16,
              "maxLength": 128
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SealEnvelopeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The sealed envelope.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/Denied"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "503": {
            "$ref": "#/components/responses/DependencyUnavailable"
          }
        }
      }
    },
    "/v1/operations/sign": {
      "post": {
        "operationId": "sign",
        "tags": [
          "operations"
        ],
        "parameters": [
          {
            "name": "X-Request-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "description": "Must exactly equal context.nonce; every value is accepted at most once.",
            "schema": {
              "type": "string",
              "minLength": 16,
              "maxLength": 128
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authorized result.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/Denied"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/InvalidOperation"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        }
      }
    },
    "/v1/operations/unwrap": {
      "post": {
        "operationId": "unwrapDataKey",
        "tags": [
          "operations"
        ],
        "parameters": [
          {
            "name": "X-Request-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "description": "Must exactly equal context.nonce; every value is accepted at most once.",
            "schema": {
              "type": "string",
              "minLength": 16,
              "maxLength": 128
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnwrapRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authorized unwrapped data key.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/Denied"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/InvalidOperation"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        }
      }
    },
    "/v1/operations/wrap": {
      "post": {
        "operationId": "wrapDataKey",
        "tags": [
          "operations"
        ],
        "parameters": [
          {
            "name": "X-Request-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "description": "Must exactly equal context.nonce; every value is accepted at most once.",
            "schema": {
              "type": "string",
              "minLength": 16,
              "maxLength": 128
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WrapRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authorized hardware-wrapped data key.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/Denied"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/InvalidOperation"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "mutualTLS": {
        "type": "mutualTLS",
        "description": "Client certificate mapped to a workload principal."
      }
    },
    "schemas": {
...
      "OperationResult": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "request_id",
          "operation_id",
          "object_id",
          "content_type",
          "result_base64"
        ],
        "properties": {
          "request_id": {
            "type": "string",
            "format": "uuid"
          },
          "operation_id": {
            "type": "string",
            "format": "uuid"
          },
          "object_id": {
            "$ref": "#/components/schemas/ObjectID"
          },
          "content_type": {
            "type": "string",
            "maxLength": 128
          },
          "result_base64": {
            "type": "string",
            "maxLength": 1400000
          }
        }
      },
...
      "UnwrapRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "object_id",
          "context",
          "format",
          "wrapped_data_key_base64"
        ],
        "properties": {
          "object_id": {
            "$ref": "#/components/schemas/ObjectID"
          },
          "context": {
            "$ref": "#/components/schemas/OperationContext"
          },
          "format": {
            "enum": [
              "sops-pgp",
              "regalia-envelope-v2"
            ]
          },
          "wrapped_data_key_base64": {
            "type": "string",
            "maxLength": 65536,
            "pattern": "^[A-Za-z0-9+/]*={0,2}$"
          },
          "envelope_aad_base64": {
            "type": "string",
            "maxLength": 16384,
            "pattern": "^[A-Za-z0-9+/]*={0,2}$"
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Malformed or invalid request.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthenticated": {
        "description": "No valid workload identity.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Denied": {
        "description": "Identity or policy denied the request.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "Conflict": {
        "description": "Idempotency/replay nonce already consumed or durable state conflict.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "InvalidOperation": {
        "description": "Operation is incompatible with object purpose or policy.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unavailable": {
        "description": "Required backend, policy, registry, audit, or state dependency is unavailable.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  }
}